Fix: [#31] 단일포스트조회 에러해결

+++ back/dipping/src/main/java/com/common/dipping/api/board/domain/dto/BoardResponse.java
@@ -17,7 +17,7 @@ public class BoardResponse {
 	private Boolean albumArt;
 	private Long likeCount;
 	private Boolean myLike;
-	private Long commentCount;
+	private int commentCount;
 	
 	public BoardResponse(Board board) {
 		this.boardId = board.getId();
